Support new SpiderMonkey API. wposix: Define int8_t compatibly with SpiderMonkey. Remove unused camera, lightenv JS interfaces. Remove most of vector JS interface. Remove some of the redundant JS string conversion functions. Remove unneeded vmem, _lodBias functions. Clean up some formatting.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@8629 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/GameView.cpp
@@ -676,7 +676,7 @@ void CGameView::Update(float DeltaTime)
 				cmpPosition->GetInterpolatedPosition2D(frameOffset, x, z, angle);
 				float height = 4.f;
 				m->ViewCamera.m_Orientation.SetIdentity();
-				m->ViewCamera.m_Orientation.RotateX(M_PI/24.f);
+				m->ViewCamera.m_Orientation.RotateX((float)M_PI/24.f);
 				m->ViewCamera.m_Orientation.RotateY(angle);
 				m->ViewCamera.m_Orientation.Translate(pos.X, pos.Y + height, pos.Z);
